refactor: extrack stack
